Remove unnecessary files & start using .env
Removed unnecessary test files that were added
as a placeholder.

Also instead of the master.key method for storing
credentials in an encrypted form in the resository,
going back to the .env method based on the following
discussion on slack:

In our previous projects, we experienced that any time
we needed to change a key in production, it can be done
very easily via heroku interface if those are implemneted
via the .env file.

The master.key method, I suppose would need a programmer to:
- create a PR  to change to that master.key file to do the same
- Wait for a reviewer to review and approve it
- Then when its merged changes show up on Heroku.

This could get cumbersome during the final stages
of the project where if due to some unforeseen
circumstances we have to change one of the API keys.